requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.16.0
transformers>=4.40.0
huggingface-hub>=0.20.0
safetensors>=0.4.0

# Data & Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# App
gradio>=4.0.0

# MedGemma (via Ollama)
ollama>=0.1.0

# Utilities
tqdm>=4.65.0
pydicom>=2.4.0
requests>=2.31.0
psutil>=5.9.0
